requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
motor[srv]>=3.6.0
pydantic[email]==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
slowapi==0.1.9
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
python-dotenv==1.0.1
